refactor: 프로젝트 폴더 구조 개선
