+++ src/components/card/Card/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CARD_NUMBER_TYPE, EXPIRATION_DATE_TYPE, PASSWORD_TYPE } from '../types';
-import { objectToString } from '../../utils/util';
-import { CARD_TYPE } from '../../utils/constants';
+import { CARD_NUMBER_TYPE, EXPIRATION_DATE_TYPE, PASSWORD_TYPE } from '../../types';
+import { objectToString } from '../../../utils/util';
+import { CARD_TYPE } from '../../../utils/constants';
 
 function Card({ cardInformation: { cardNumber, expirationDate, ownerName, cardType } }) {
   return (
