+++ src/components/input/InputContainer/TypeInputContainer/SecurityCode/SecurityCode.stories.js
